requirements.txt
fastapi
uvicorn
boto3
python-multipart
python-dotenv
openai
psycopg2-binary
sqlalchemy
mangum
python-jose[cryptography]
requests